refactor: infra 모듈 infrastructure로 네이밍 통일
