+++ backend/pcloud-api/src/main/java/com/api/global/config/moduleutils/ComponentScanConfig.java
@@ -4,6 +4,6 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan(value = {"com.common", "com.domain", "com.infra"})
+@ComponentScan(value = {"com.common", "com.domain", "com.infrastructure"})
 public class ComponentScanConfig {
 }
