+++ backend/pcloud-infrastructure/src/main/java/com/infrastructure/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.infra.config;
+package com.infrastructure.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.jsontype.BasicPolymorphicTypeValidator;
