+++ backend/pcloud-infrastructure/src/test/java/com/infrastructure/PCloudInfraApplicationTest.java
@@ -1,4 +1,4 @@
-package com.infra;
+package com.infrastructure;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
